pyproject.toml
# do not add any other tools within the tables belonging to tool.poetry. add them at the end.
# see blake at the bottom

[tool.poetry]
name = "dlt-verified-sources"
version = "0.1.0"
description = "Initial set of dlt sources with demo pipelines installable with `dlt init` command"
authors = ["Various Authors <team@dlthub.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sources"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
dlt = {version = "0.4.7", allow-prereleases = true, extras = ["redshift", "bigquery", "postgres", "duckdb", "s3", "gs"]}
graphlib-backport = {version = "*", python = "<3.9"}

[tool.poetry.group.dev.dependencies]
mypy = "1.6.1"
flake8 = "^6.0.0"
pytest = "^7.2.0"
bandit = "^1.7.4"
flake8-bugbear = "^22.12.6"
flake8-tidy-imports = "^4.8.0"
flake8-encodings = "^0.5.0.post1"
flake8-builtins = "^2.1.0"
types-requests = "^2.28.11.7"
mimesis = "^7.0.0"
requests-mock = "^1.10.0"
pandas-stubs = "^2.0.0"
black = "^23.3.0"
pypdf2 = "^3.0.1"
greenlet = "<3.0.0"
confluent-kafka = "^2.3.0"
pytest-mock = "^3.12.0"
twisted = "22.10.0"
pytest-forked = "^1.6.0"

[tool.poetry.group.sql_database.dependencies]
sqlalchemy = ">=1.4"
pymysql = "^1.0.3"

[tool.poetry.group.google_sheets.dependencies]
google-api-python-client = "^2.78.0"

[tool.poetry.group.google_analytics.dependencies]
google-analytics-data = "^0.16.2"
google-api-python-client = "^2.86.0"
google-auth-oauthlib = "^1.0.0"
requests-oauthlib = "^1.3.1"

[tool.poetry.group.stripe_analytics.dependencies]
pandas = "^2.0.0"
stripe = "^5.0.0"
types-stripe = "^3.5.2.14"

[tool.poetry.group.asana_dlt.dependencies]
asana = "^3.2.1"

[tool.poetry.group.facebook_ads.dependencies]
facebook-business = "^17.0.2"

[tool.poetry.group.google_ads.dependencies]
google-ads = "^21.1.0"

[tool.poetry.group.salesforce.dependencies]
simple-salesforce = "^1.12.4"

[tool.poetry.group.unstructured_data_lint.dependencies]
langchain = "^0.0.219"
openai = "^0.27.8"

[tool.poetry.group.unstructured_data.dependencies]
langchain = "^0.0.219"
unstructured = "^0.7.10"
openai = "^0.27.8"
chromadb = "^0.3.26"
tiktoken = "^0.4.0"

[tool.poetry.group.mongodb.dependencies]
pymongo = "^4.3.3"

[tool.poetry.group.airtable.dependencies]
pyairtable = "^2.1.0.post1"


[tool.poetry.group.filesystem.dependencies]
adlfs = ">=2023.9.0"


[tool.poetry.group.scrapy.dependencies]
scrapy = "^2.11.0"
twisted = "22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '.*py$'